--- src/packages/core/collection/collection-picker-modal/index.ts
@@ -1,1 +1,0 @@
-export * from './collection-picker-modal.context.js';
