+++ src/packages/core/collection/index.ts
@@ -5,7 +5,7 @@ import './components/index.js';
 export * from './default/collection-default.element.js';
 export * from './collection.element.js';
 export * from './components/index.js';
-export * from './collection-picker-modal/index.js';
+export * from './collection-item-picker-modal/index.js';
 
 export * from './default/collection-default.context.js';
 export * from './default/collection-default.context-token.js';
